requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
garmin-fit-sdk==21.141.0
python-multipart==0.0.19
httpx==0.28.1
pillow==11.1.0
jinja2==3.1.5
python-dotenv==1.0.1
pydantic-settings==2.7.1
